Upgrade capital to allow storing larger values of money

+++ db/src/schema.rs
@@ -18,7 +18,7 @@ table! {
     user (uuid) {
         uuid -> Text,
         name -> Text,
-        capital -> Integer,
+        capital -> BigInt,
     }
 }
